+++ src/core/utils.zig
@@ -54,7 +54,7 @@ pub fn printUsage(io: std.Io) !void {
 }
 
 pub fn printVersion(io: std.Io) !void {
-    const version = "v1.2.1";
+    const version = "v1.2.2";
     var buf: [64]u8 = undefined;
     var w = stdoutWriter(io, &buf);
     try w.interface.print("{s}\n", .{version});